linux-gen: pktio: socket: move common code away from socket pktios
